+++ common/database/src/main/java/ru/maplyb/unitmanagerlib/common/database/entity/PositionEntity.kt
@@ -4,7 +4,7 @@ import androidx.room.Entity
 
 @Entity(primaryKeys = ["x", "y", "name"])
 data class PositionEntity(
-    val x: Double,
-    val y: Double,
+    val x: Int,
+    val y: Int,
     val name: String
 )
